backend/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
python-multipart==0.0.20
tensorflow-cpu==2.18.0
numpy==1.26.4
pillow==11.0.0
opencv-python-headless==4.10.0.84